chore(test):add test framework [Fix #167730740]
