+++ package.json
@@ -4,7 +4,7 @@
   "description": "A Social platform for the creative at heart",
   "main": "index.js",
   "scripts": {
-    "test": "cross-env NODE_ENV=development nyc mocha --require @babel/register --require @babel/polyfill ./src/__tests__/*/*.js --exit",
+    "test": "cross-env NODE_ENV=development nyc mocha --require @babel/register --require @babel/polyfill ./src/__tests__/*.js --exit",
     "coverage": "nyc report --reporter=text",
     "coveralls": "nyc report --reporter=text-lcov | coveralls",
     "lint": "eslint \"**/*.js\" --ignore-pattern node_modules/ --fix",
