+++ src/__tests__/mail.js
@@ -1,0 +1,34 @@
+import chai from 'chai';
+import chaiHttp from 'chai-http';
+import sinon from 'sinon';
+import sinonChai from 'sinon-chai';
+import app from '../index';
+
+
+chai.use(chaiHttp);
+chai.use(sinonChai);
+
+const { expect } = chai;
+
+let request;
+
+describe('EMAIL ROUTE', () => {
+  before(async () => {
+    request = chai.request(app).keepOpen();
+  });
+
+  afterEach(() => sinon.restore());
+
+  describe('EMAIL VERIFICATION ROUTE', () => {
+    it('should have a status of 200 when message is sent succesfully', async () => {
+      const body = {
+        email: 'akp.ani@yahoo.com',
+        firstName: 'Aniefiok',
+        lastName: 'Akpan'
+      };
+      const response = await request.post('/api/email-test').send(body);
+      expect(response.body.status).to.equal(200);
+      expect(response.body).to.be.a('object');
+    }).timeout(0);
+  });
+});
